requirements.txt
fastapi==0.84.0
uvicorn==0.18.3
requests
pyyaml
